+++ src/main/java/com/github/danielfabela/partpicker/io/Dao.java
@@ -1,0 +1,8 @@
+package com.github.danielfabela.partpicker.io;
+
+import java.util.List;
+
+public interface Dao<E> {
+    List<E> readAll();
+
+}